requirements.txt
# GW 자동화 프로젝트 의존성

# AI (Gemini)
google-genai>=1.0.0

# 웹 프레임워크
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# 환경 변수
python-dotenv>=1.0.0

# 브라우저 자동화 & HTTP
playwright>=1.49.0
requests>=2.32.0
httpx>=0.27.0

# 다중 사용자 인증
cryptography>=42.0.0
PyJWT>=2.8.0

# 텔레그램 연동
python-telegram-bot>=21.0

# Google Cloud APIs
google-cloud-speech>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0
